rename files with snake_case
